# SMM-Dashboard/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SMM Panel API - Perfect Follow
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get user orders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Insufficient balance or bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /payments:
    get:
      operationId: getPayments
      tags: [payments]
      summary: Get user payments
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a manual payment request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/upload-receipt:
    post:
      operationId: uploadReceipt
      tags: [payments]
      summary: Upload payment receipt image (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadReceiptBody"
      responses:
        "200":
          description: Receipt uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadReceiptResponse"
...
  /admin/users/{id}/balance:
    put:
      operationId: adminUpdateUserBalance
      tags: [admin]
      summary: Admin - Update user balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBalanceBody"
      responses:
        "200":
          description: Balance updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/orders/{id}/status:
    put:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Admin - Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /admin/payments/{id}/approve:
    put:
      operationId: adminApprovePayment
      tags: [admin]
      summary: Admin - Approve a payment and credit balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /admin/payments/{id}/reject:
    put:
      operationId: adminRejectPayment
      tags: [admin]
      summary: Admin - Reject a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        balance:
          type: number
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - balance
        - role
        - createdAt
...
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        platformName:
          type: string
        link:
          type: string
        quantity:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, processing, completed, cancelled, failed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - serviceId
        - serviceName
        - platformName
        - link
        - quantity
        - totalPrice
        - status
        - createdAt
    CreateOrderBody:
      type: object
      properties:
        serviceId:
          type: integer
        link:
          type: string
        quantity:
          type: integer
      required:
        - serviceId
        - link
        - quantity
    Payment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: number
        method:
          type: string
          enum: [zaincash, asiahawala, stripe]
        status:
          type: string
          enum: [pending, approved, rejected]
        transactionId:
          type: string
        receiptUrl:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - amount
        - method
        - status
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [zaincash, asiahawala, stripe]
        transactionId:
          type: string
        receiptUrl:
          type: string
        notes:
          type: string
      required:
        - amount
        - method
    UploadReceiptBody:
      type: object
      properties:
        imageData:
          type: string
          description: Base64 encoded image data
        fileName:
          type: string
      required:
        - imageData
        - fileName
    UploadReceiptResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...
    UpdateBalanceBody:
      type: object
      properties:
        balance:
          type: number
        operation:
          type: string
          enum: [set, add, subtract]
      required:
        - balance
        - operation
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, completed, cancelled, failed]
      required:
        - status
...